DTNN-1128: Make indices into the MIB array uint16s

+++ aa/as/inc/bplib_as.h
@@ -247,7 +247,7 @@ void BPLib_AS_Decrement(BPLib_EID_t EID, BPLib_AS_Counter_t Counter, uint32_t Am
  *            The index into the source MIB counter array must be known ahead of time. This can
  *            be determined by looking at the source EIDs inside the MIB array, nested within the
  *            source MIB counters payload of the source MIB counters HK packet
- * \param[in] MibArrayIndex (uint8_t) Index into BPLib_SourceMibCountersHkTlm_Payload_t::MibArray
+ * \param[in] MibArrayIndex (uint16_t) Index into BPLib_SourceMibCountersHkTlm_Payload_t::MibArray
  * \param[in] Counter       (BPLib_AS_Counter_t) Counter to reset
  * \return    Execution status
  * \retval    BPLIB_AS_UNKNOWN_NODE_CNTR: Counter is outside node counters' range
@@ -267,7 +267,7 @@ BPLib_Status_t BPLib_AS_ResetCounter(uint16_t MibArrayIndex, BPLib_AS_Counter_t 
  * \note      The index into the source MIB counter array must be known ahead of time. This can
  *            be determined by looking at the source EIDs inside the MIB array, nested within the
  *            source MIB counters payload of the source MIB counters HK packet
- * \param[in] MibArrayIndex (uint8_t) Index into BPLib_SourceMibCountersHkTlm_Payload_t::MibArray
+ * \param[in] MibArrayIndex (uint16_t) Index into BPLib_SourceMibCountersHkTlm_Payload_t::MibArray
  * \return    Execution status
  * \retval    BPLIB_AS_INVALID_MIB_INDEX: Given MIB counter array index is out of range
  * \retval    BPLIB_SUCCESS: Successful execution
@@ -276,7 +276,7 @@ BPLib_Status_t BPLib_AS_ResetCounter(uint16_t MibArrayIndex, BPLib_AS_Counter_t 
  * \ref       BPLib_AS_LockCounters [BPLib_AS_LockCounters()]
  * \ref       BPLib_AS_UnlockCounters [BPLib_AS_UnlockCounters()]
  */
-BPLib_Status_t BPLib_AS_ResetSourceCounters(uint8_t MibArrayIndex);
+BPLib_Status_t BPLib_AS_ResetSourceCounters(uint16_t MibArrayIndex);
 
 /**
  * \brief     Set to zero all resettable node MIB counters associated with bundles
@@ -294,7 +294,7 @@ void BPLib_AS_ResetBundleCounters(void);
  * \brief     Set to zero all resettable MIB error counters
  * \details   See function body and reference the BPLib_AS_Counter_t struct to see which counters are reset
  * \note      Directly sets error counters in node and source MIB counter payloads to 0
- * \param[in] MibArrayIndex (uint8_t) Index into BPLib_SourceMibCountersHkTlm_Payload_t::MibArray
+ * \param[in] MibArrayIndex (uint16_t) Index into BPLib_SourceMibCountersHkTlm_Payload_t::MibArray
  * \return    Execution status
  * \retval    BPLIB_AS_INVALID_MIB_INDEX: Index into source MIB counters array is out of range
  * \retval    BPLIB_SUCCESS: Successful execution
@@ -304,7 +304,7 @@ void BPLib_AS_ResetBundleCounters(void);
  * \ref       BPLib_AS_LockCounters [BPLib_AS_LockCounters()]
  * \ref       BPLib_AS_UnlockCounters [BPLib_AS_UnlockCounters()]
  */
-BPLib_Status_t BPLib_AS_ResetErrorCounters(uint8_t MibArrayIndex);
+BPLib_Status_t BPLib_AS_ResetErrorCounters(uint16_t MibArrayIndex);
 
 /**
  * \brief     Set every counter value in the source and node counter packets to zero
